Create Main code
As shows onevariate and multi-variate analysis there are no anormalities or outlieres in the data set, those points outside the cogs boxplot are expensive products and are a small group, thus, we will considere the data as it stands. This was double checked with the z-score and it shows that those points are within the +/- 3 deviations from the mean.