Create Models
The linear regression model has the lowest RMSE, thus is recommended for further exploration. Noting that if more features are added a new analysis should be made to double check the impact of those features. The selected model is Linear Regression model becasue it has the lowest RMSE. 

In general, the dataset is clean and ready to analyze, some feature transformation were made and some columns deleted due to their lack of use.  Linear regression model had the lowest RMSE, as regularization was added the performance of the model decrease with both Lasso and Ridge, this is due to the lack of impactful features. The fact that LR model had a R around 1 is rare so a more in deep analys once new features are gathered is needed. 

The correlation among the features is too low and too high in two of them. Therefore, is necessary to gather more information to add new features and improve the interpretability of the model as well as enhance the performance of the predictive models.